Update main.py
added replace/restore all and file naming, updating driver.xml

+++ main.py
@@ -4,10 +4,11 @@ from tkinter import filedialog
 from PIL import ImageTk, Image
 import os
 import shutil
+import xml.dom.minidom as md
 
 root = tk.Tk()
-root.geometry("290x390")
-root.title("Control4 Icon Swapper")
+root.geometry('290x410')
+root.title('Control4 Icon Swapper')
 root.resizable(False, False)
 
 entry_font = 'Helvetica'
@@ -15,39 +16,43 @@ entry_font = 'Helvetica'
 global icon_objects
 global current_icon
 global replacement_image
+global driver_name
 
 driver_selected = False
 replacement_selected = False
 
 
 class Icon:
-    def __init__(self, index: int, icon_family: str, icon_names: list, icon_paths: list):
-        self.index = index
+    def __init__(self, icon_family: str, icon_names: list, icon_paths: list, do_init=True):
         self.icon_family = icon_family
         self.icon_sizes = []
         self.icon_names = []
         self.icon_paths = []
 
-        for path in icon_paths:
-            if str(path).endswith('.orig'):
-                continue
-            else:
-                self.icon_paths.append(path)
+        if do_init:
+            for path in icon_paths:
+                if str(path).endswith('.orig'):
+                    continue
+                else:
+                    self.icon_paths.append(path)
 
-        for name in icon_names:
-            if str(name).endswith('.orig'):
-                continue
-            else:
-                self.icon_names.append(name)
-            temp_size = ''
-            for l in range(len(name)):
-                if name[l] == '.':
-                    break
-                elif name[l] == '_':
-                    temp_size += name[l + 1]
-                elif temp_size != '' and name[l - 1] != '_':
-                    temp_size += name[l]
-            self.icon_sizes.append(int(temp_size))
+            for name in icon_names:
+                if str(name).endswith('.orig'):
+                    continue
+                else:
+                    self.icon_names.append(name)
+                temp_size = ''
+                for l in range(len(name)):
+                    if name[l] == '.':
+                        break
+                    elif name[l] == '_':
+                        temp_size += name[l + 1]
+                    elif temp_size != '' and name[l - 1] != '_':
+                        temp_size += name[l]
+                self.icon_sizes.append(int(temp_size))
+        else:
+            self.icon_names = icon_names
+            self.icon_paths = icon_paths
 
     def show_icon(self):
         global current_icon
@@ -55,6 +60,7 @@ class Icon:
 
         if os.path.isfile(self.icon_paths[0] + '.orig'):
             restore_button['state'] = ACTIVE
+            restore_all_button['state'] = ACTIVE
         else:
             restore_button['state'] = DISABLED
         icon_image = Image.open(self.icon_paths[0])
@@ -66,23 +72,18 @@ class Icon:
         icon_image_label2.image = icon
         icon_image_label2.grid(row=2, column=1)
 
-        if len(icon_objects) == 1:
-            prev_icon_button['state'] = DISABLED
-            next_icon_button['state'] = DISABLED
-        else:
-            prev_icon_button['state'] = ACTIVE
-            next_icon_button['state'] = ACTIVE
-
 
 def upload_c4z():
     global icon_objects
     global current_icon
     global driver_selected
     global replacement_selected
+    global driver_name
 
     icon_objects = []
     filename = filedialog.askopenfilename(filetypes=[("Control4 Drivers", "*.c4z")])
     icon_dir = temp_dir + '/driver/www/icons/device'
+    composer_icon_dir = temp_dir + '/driver/www/icons'
 
     file_entry_field['state'] = NORMAL
     file_entry_field.delete(0, 'end')
@@ -116,12 +117,18 @@ def upload_c4z():
                 children_names.append(icons_list[j])
                 children_paths.append(icon_dir + '/' + icons_list[j])
 
-            icon_objects.append(Icon(i, icon_families[i], children_names, children_paths))
+            icon_objects.append(Icon(icon_families[i], children_names, children_paths))
         file_entry_field['state'] = NORMAL
         file_entry_field.insert(0, filename)
 
         icon_objects[0].show_icon()
         current_icon = 1
+
+        icon_objects.append(Icon('device', ['device_lg.png', 'device_sm.png'], [composer_icon_dir + '/device_lg.png',
+                                                                                composer_icon_dir + '/device_sm.png'],
+                                 do_init=False))
+        icon_objects[len(icon_objects) - 1].icon_sizes = [32, 16]
+
         icon_label.config(text='icon: ' + str(current_icon) + ' of ' + str(len(icon_objects)))
         icon_name_label.config(text='icon name: ' + icon_objects[current_icon - 1].icon_names[0])
 
@@ -139,7 +146,30 @@ def upload_c4z():
     if driver_selected:
         if replacement_selected:
             replace_icon_button['state'] = ACTIVE
+            replace_all_button['state'] = ACTIVE
         export_button['state'] = ACTIVE
+
+        driver_xml = open(temp_dir + '/driver/driver.xml', encoding='utf-8')
+        driver_xml_str = driver_xml.readlines()
+        temp = []
+        for string in driver_xml_str:
+            temp.append(str(string.encode('ascii', errors='replace')))
+        driver_xml_str = temp
+
+        driver_name = ''
+        for string in driver_xml_str:
+            if '<name>' in string:
+                result = re.search('<name>(.*)</name>', string)
+                driver_name = result.group(1).replace(' ', '')
+                break
+
+        driver_name_entry.insert(0, driver_name)
+    if len(icon_objects) == 1:
+        prev_icon_button['state'] = DISABLED
+        next_icon_button['state'] = DISABLED
+    elif len(icon_objects) > 1:
+        prev_icon_button['state'] = ACTIVE
+        next_icon_button['state'] = ACTIVE
 
 
 def upload_replacement():
@@ -166,6 +196,7 @@ def upload_replacement():
 
     if driver_selected:
         replace_icon_button['state'] = ACTIVE
+        replace_all_button['state'] = ACTIVE
     replacement_selected = True
 
 
@@ -187,16 +218,49 @@ def replace_icon():
 
     replacement_dir = temp_dir + '/replacement_icon/'
     icon_dir = temp_dir + '/driver/www/icons/device/'
+    comp_dir = temp_dir + '/driver/www/icons/'
 
     for icon in os.listdir(replacement_dir):
-        if not os.path.isfile(icon_dir + icon + '.orig'):
+        if icon == 'device_lg.png' or icon == 'device_sm.png':
+            if not os.path.isfile(comp_dir + icon + '.orig'):
+                os.rename(comp_dir + icon, comp_dir + icon + '.orig')
+                shutil.copy(replacement_dir + icon, comp_dir)
+                os.remove(replacement_dir + icon)
+        elif not os.path.isfile(icon_dir + icon + '.orig'):
             os.rename(icon_dir + icon, icon_dir + icon + '.orig')
-        shutil.copy(replacement_dir + icon, icon_dir)
-        os.remove(replacement_dir + icon)
+            shutil.copy(replacement_dir + icon, icon_dir)
+            os.remove(replacement_dir + icon)
 
     icon_objects[current_icon - 1].show_icon()
 
     restore_button['state'] = ACTIVE
+    restore_all_button['state'] = ACTIVE
+
+
+def replace_all():
+    global current_icon
+
+    current_icon = 1
+
+    for i in range(0, len(icon_objects)):
+        replace_icon()
+        next_icon()
+
+    icon_objects[current_icon - 1].show_icon()
+
+
+def restore_all():
+    global current_icon
+    global icon_objects
+
+    current_icon = 1
+
+    for i in range(0, len(icon_objects)):
+        restore_icon()
+        next_icon()
+
+    icon_objects[current_icon - 1].show_icon()
+    restore_all_button['state'] = DISABLED
 
 
 def restore_icon():
@@ -211,6 +275,10 @@ def restore_icon():
 
     icon_objects[current_icon - 1].show_icon()
     restore_button['state'] = DISABLED
+    restore_all_button['state'] = DISABLED
+    for icon in icon_objects:
+        if os.path.isfile(icon.icon_paths[0] + '.orig'):
+            restore_all_button['state'] = ACTIVE
 
 
 def next_icon():
@@ -246,24 +314,51 @@ def prev_icon():
 # noinspection PyGlobalUndefined
 def export_driver():
     global pop
+    global temp_dir
+    global driver_name
+
+    driver_name = driver_name_entry.get().replace(' ', '')
+    temp = ''
+    for letter in driver_name:
+        if str(letter).isalnum() or str(letter) == '_' or str(letter) == '-':
+            temp += str(letter)
+    driver_name = temp
+
+    driver_name_entry.delete(0, 'end')
+    driver_name_entry.insert(0, driver_name)
+
+    # replace xml data
+    driver_xml = md.parse(temp_dir + '/driver/driver.xml')
+    driver_xml.getElementsByTagName('name')[0].childNodes[0].nodeValue = driver_name
+    for i in range(driver_xml.getElementsByTagName('Icon').length):
+        temp_str = driver_xml.getElementsByTagName('Icon')[i].childNodes[0].nodeValue
+        result = re.search('driver/(.*)/icons', temp_str)
+        result_str = result.group(1)
+        temp_str = temp_str.replace(result_str, driver_name)
+        driver_xml.getElementsByTagName('Icon')[i].childNodes[0].nodeValue = temp_str
+
+    with open(temp_dir + '/driver/driver.xml', "w") as fs:
+        fs.write(driver_xml.toxml())
+        fs.close()
 
     def confirm_write():
-        if os.path.isfile(cur_dir + '/new_driver.c4z'):
-            os.remove(cur_dir + '/new_driver.c4z')
-        if os.path.isfile(cur_dir + '/new_driver.zip'):
-            os.remove(cur_dir + '/new_driver.zip')
-        shutil.make_archive('new_driver', 'zip', temp_dir + '/driver')
-        base_path = os.path.splitext(cur_dir + '/new_driver.zip')[0]
-        os.rename(cur_dir + '/new_driver.zip', base_path + '.c4z')
+        if os.path.isfile(cur_dir + '/' + driver_name + '.c4z'):
+            os.remove(cur_dir + '/' + driver_name + '.c4z')
+        if os.path.isfile(cur_dir + '/' + driver_name + '.zip'):
+            os.remove(cur_dir + '/' + driver_name + '.zip')
+        shutil.make_archive(driver_name, 'zip', temp_dir + '/driver')
+        base_path = os.path.splitext(cur_dir + '/' + driver_name + '.zip')[0]
+        os.rename(cur_dir + '/' + driver_name + '.zip', base_path + '.c4z')
 
         pop.destroy()
 
-    if os.path.isfile(cur_dir + '/new_driver.c4z') or os.path.isfile(cur_dir + '/new_driver.zip'):
+    if os.path.isfile(cur_dir + '/' + driver_name + '.c4z') or os.path.isfile(cur_dir + '/' + driver_name + '.zip'):
         pop = Toplevel(root)
         pop.title('Overwrite')
         pop.geometry('239x50')
         pop.grab_set()
         pop.transient(root)
+        pop.resizable(False, False)
 
         confirm_label = Label(pop, text='Would you like to overwrite the existing file?')
         confirm_label.grid(row=0, column=0, columnspan=2)
@@ -274,9 +369,9 @@ def export_driver():
         yes_button = tk.Button(pop, text='Yes', width='10', command=confirm_write)
         yes_button.grid(row=1, column=1)
     else:
-        shutil.make_archive('new_driver', 'zip', temp_dir + '/driver')
-        base = os.path.splitext(cur_dir + '/new_driver.zip')[0]
-        os.rename(cur_dir + '/new_driver.zip', base + '.c4z')
+        shutil.make_archive(driver_name, 'zip', temp_dir + '/driver')
+        base = os.path.splitext(cur_dir + '/' + driver_name + '.zip')[0]
+        os.rename(cur_dir + '/' + driver_name + '.zip', base + '.c4z')
 
 
 cur_dir = os.getcwd()
@@ -305,10 +400,10 @@ open_file_button = tk.Button(root, text='Open', command=upload_c4z)
 open_file_button.grid(row=1, column=2, sticky='W')
 
 restore_button = tk.Button(root, text='Restore Original Icon', command=restore_icon)
-restore_button.grid(row=2, column=1, columnspan=2, sticky='E', padx=5)
+restore_button.grid(row=2, column=1, columnspan=2, sticky='E', padx=10)
 restore_button['state'] = DISABLED
 
-prev_icon_button = tk.Button(root, text='Prev', command=prev_icon)
+prev_icon_button = tk.Button(root, text='Prev', command=prev_icon, width=5)
 prev_icon_button.grid(row=3, column=1, sticky='W')
 prev_icon_button['state'] = DISABLED
 
@@ -318,7 +413,7 @@ icon_label.grid(row=3, column=1)
 icon_name_label = tk.Label(root, text='icon name:')
 icon_name_label.grid(row=4, column=1, columnspan=3, sticky='W')
 
-next_icon_button = tk.Button(root, text='Next', command=next_icon)
+next_icon_button = tk.Button(root, text='Next', command=next_icon, width=5)
 next_icon_button.grid(row=3, column=1, sticky='E')
 next_icon_button['state'] = DISABLED
 
@@ -335,16 +430,30 @@ replacement_entry_field['state'] = 'readonly'
 open_file2_button = tk.Button(root, text='Open', command=upload_replacement)
 open_file2_button.grid(row=7, column=2, sticky='W')
 
-replace_icon_button = tk.Button(root, text='Replace current icon', command=replace_icon)
+replace_icon_button = tk.Button(root, text='Replace Current Icon', command=replace_icon)
 replace_icon_button.grid(row=9, column=1)
 replace_icon_button['state'] = DISABLED
+
+replace_all_button = tk.Button(root, text='Replace All', command=replace_all)
+replace_all_button.grid(row=9, column=2, sticky='W')
+replace_all_button['state'] = DISABLED
+
+restore_all_button = tk.Button(root, text='Restore All', command=restore_all)
+restore_all_button.grid(row=8, column=2, sticky='W')
+restore_all_button['state'] = DISABLED
 
 divider_label2 = tk.Label(root, text='---------------------------------------------------------')
 divider_label2.grid(row=10, column=0, columnspan=3, sticky='E')
 
 export_button = tk.Button(root, text='Export Control4 Driver', width=35, command=export_driver)
-export_button.grid(row=11, column=0, columnspan=3, padx=5)
+export_button.grid(row=14, column=0, columnspan=3, padx=5, pady=5)
 export_button['state'] = DISABLED
+
+driver_name_label = Label(root, text='Driver Name:')
+driver_name_label.grid(row=11, column=1, sticky='W')
+
+driver_name_entry = Entry(root, text='', width=25)
+driver_name_entry.grid(row=11, column=1, columnspan=2, sticky='E', padx=20)
 
 root.mainloop()
